# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VyOS API",
    "description": "REST proxy for one or more VyOS router devices. Exposes CRUD operations over interfaces, VRFs, VLANs, firewall policies, and firewall address groups. Each device is identified by the name supplied in the VYOS_HOSTS environment variable.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/devices/{device_id}/firewall/policies": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" }
      ],
      "get": {
        "tags": ["firewall"],
        "summary": "List all firewall policies",
        "description": "Returns all IPv4 named firewall policies. Rules are not included in the list view; use GET on a specific policy to retrieve rules.",
        "operationId": "listPolicies",
        "responses": {
          "200": {
            "description": "Policy list (rules omitted)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PolicyInfo" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "post": {
        "tags": ["firewall"],
        "summary": "Create a firewall policy",
        "operationId": "createPolicy",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePolicyRequest" },
              "example": {
                "name": "LAN-IN",
                "default_action": "drop",
                "description": "inbound-lan-traffic"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policy created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PolicyInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/firewall/policies/{policy}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/policy" }
      ],
      "get": {
        "tags": ["firewall"],
        "summary": "Get a firewall policy",
        "description": "Returns the policy including all of its rules, keyed by rule number.",
        "operationId": "getPolicy",
        "responses": {
          "200": {
            "description": "Policy details including rules",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PolicyInfo" },
                "example": {
                  "name": "LAN-IN",
                  "default_action": "drop",
                  "description": "inbound-lan-traffic",
                  "rules": {
                    "10": { "action": "accept", "source": "10.0.0.0/8", "destination": "", "description": "" },
                    "20": { "action": "drop",   "source": "",            "destination": "", "description": "" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "put": {
        "tags": ["firewall"],
        "summary": "Update a firewall policy",
        "description": "Updates `default_action` and/or `description`. Omit fields that should not change. Rules are managed via the `/rules` sub-resource.",
        "operationId": "updatePolicy",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePolicyRequest" },
              "example": { "default_action": "accept" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PolicyInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "delete": {
        "tags": ["firewall"],
        "summary": "Delete a firewall policy",
        "description": "Removes the entire policy including all its rules.",
        "operationId": "deletePolicy",
        "responses": {
          "204": { "description": "Policy deleted" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/firewall/policies/{policy}/rules": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/policy" }
      ],
      "post": {
        "tags": ["firewall"],
        "summary": "Add a rule to a policy",
        "description": "Adds a numbered rule to an existing policy. `rule_id` must be a positive integer (VyOS convention: multiples of 10). If the rule already exists it is overwritten.",
        "operationId": "addRule",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddRuleRequest" },
              "example": {
                "rule_id": 10,
                "action": "accept",
                "source": "10.0.0.0/8",
                "description": "allow-rfc1918"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rule added",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RuleResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/firewall/policies/{policy}/rules/{rule_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/policy" },
        { "$ref": "#/components/parameters/rule_id" }
      ],
      "delete": {
        "tags": ["firewall"],
        "summary": "Delete a rule",
        "operationId": "deleteRule",
        "responses": {
          "204": { "description": "Rule deleted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "device_id": {
        "name": "device_id",
        "in": "path",
        "required": true,
        "description": "Device name as registered in VYOS_HOSTS",
        "schema": { "type": "string", "example": "router1" }
      },
...
      "policy": {
        "name": "policy",
        "in": "path",
        "required": true,
        "description": "Firewall policy name",
        "schema": { "type": "string", "example": "LAN-IN" }
      },
      "rule_id": {
        "name": "rule_id",
        "in": "path",
        "required": true,
        "description": "Rule number (positive integer, VyOS convention: multiples of 10)",
        "schema": { "type": "integer", "minimum": 1, "example": 10 }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "device not found: router99" }
        }
      },
...
      "RuleInfo": {
        "type": "object",
        "required": ["action"],
        "properties": {
          "action":      { "type": "string", "enum": ["accept", "drop", "reject"], "example": "accept" },
          "source":      { "type": "string", "description": "Source IP or CIDR", "example": "10.0.0.0/8" },
          "destination": { "type": "string", "description": "Destination IP or CIDR", "example": "" },
          "description": { "type": "string", "example": "allow-rfc1918" }
        }
      },
      "PolicyInfo": {
        "type": "object",
        "required": ["name", "default_action"],
        "properties": {
          "name":           { "type": "string", "example": "LAN-IN" },
          "default_action": { "type": "string", "enum": ["accept", "drop", "reject"], "example": "drop" },
          "description":    { "type": "string", "example": "inbound-lan-traffic" },
          "rules": {
            "type": "object",
            "description": "Map of rule number (string) to rule details. Present on GET, omitted from list responses.",
            "additionalProperties": { "$ref": "#/components/schemas/RuleInfo" },
            "example": {
              "10": { "action": "accept", "source": "10.0.0.0/8", "destination": "", "description": "" }
            }
          }
        }
      },
      "CreatePolicyRequest": {
        "type": "object",
        "required": ["name", "default_action"],
        "properties": {
          "name":           { "type": "string", "example": "LAN-IN" },
          "default_action": { "type": "string", "enum": ["accept", "drop", "reject"], "example": "drop" },
          "description":    { "type": "string", "example": "inbound-lan-traffic" }
        }
      },
      "UpdatePolicyRequest": {
        "type": "object",
        "properties": {
          "default_action": { "type": "string", "enum": ["accept", "drop", "reject"], "example": "accept" },
          "description":    { "type": "string", "example": "updated-description" }
        }
      },
      "AddRuleRequest": {
        "type": "object",
        "required": ["rule_id", "action"],
        "properties": {
          "rule_id":     { "type": "integer", "minimum": 1, "description": "Rule number (VyOS convention: multiples of 10)", "example": 10 },
          "action":      { "type": "string", "enum": ["accept", "drop", "reject"], "example": "accept" },
          "source":      { "type": "string", "description": "Source IP or CIDR to match", "example": "10.0.0.0/8" },
          "destination": { "type": "string", "description": "Destination IP or CIDR to match", "example": "" },
          "description": { "type": "string", "example": "allow-rfc1918" }
        }
      },
      "RuleResponse": {
        "type": "object",
        "required": ["policy", "rule_id", "action"],
        "properties": {
          "policy":      { "type": "string", "example": "LAN-IN" },
          "rule_id":     { "type": "integer", "example": 10 },
          "action":      { "type": "string", "example": "accept" },
          "source":      { "type": "string", "example": "10.0.0.0/8" },
          "destination": { "type": "string", "example": "" },
          "description": { "type": "string", "example": "allow-rfc1918" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Missing or invalid request fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface, type, and address are required" }
          }
        }
      },
      "NotFound": {
        "description": "Device or resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface not found" }
          }
        }
      },
      "DeviceNotFound": {
        "description": "The requested device_id is not registered",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device not found: router99" }
          }
        }
      },
      "DeviceRejected": {
        "description": "The VyOS device rejected the operation (invalid config, constraint violation, etc.)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device rejected operation: Configuration path already exists" }
          }
        }
      },
      "DeviceError": {
        "description": "Could not communicate with the VyOS device (network error, TLS failure, timeout)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device communication error: dial tcp 192.168.1.1:443: connection refused" }
          }
        }
      }
    }
  }
}